+++ integration-tests/tests/test_vectors.rs
@@ -1,7 +1,7 @@
 //! Holds code used to build test vector data for unit tests. These tests should not be run by default or in CI.
 
 use anyhow::Context;
-use core2::io::{self, Read, Write};
+use corez::io::{self, Read, Write};
 use futures::TryFutureExt as _;
 use std::fs;
 use std::fs::File;
@@ -10,6 +10,7 @@ use std::io::BufWriter;
 use std::path::Path;
 use std::sync::Arc;
 use tower::{Service, ServiceExt as _};
+use wire_serialized_transaction_test_data::transactions::get_test_vectors;
 use zaino_fetch::chain::transaction::FullTransaction;
 use zaino_fetch::chain::utils::ParseFromSlice;
 use zaino_state::read_u32_le;
@@ -22,7 +23,6 @@ use zaino_state::StateService;
 use zaino_state::ZcashIndexer;
 use zaino_state::{ChainWork, IndexedBlock};
 use zaino_testutils::from_inputs;
-use wire_serialized_transaction_test_data::transactions::get_test_vectors;
 use zaino_testutils::{TestManager, ValidatorKind};
 use zcash_local_net::validator::zebrad::Zebrad;
 use zebra_chain::serialization::{ZcashDeserialize, ZcashSerialize};